extract composed form fields
