+++ src/main/app/src/app/formulieren/taken/model/abstract-task-form.spec.ts
@@ -24,7 +24,8 @@ import { RouterModule } from "@angular/router";
 import { TranslateModule } from "@ngx-translate/core";
 import { of } from "rxjs";
 import { fromPartial } from "../../../../test-helpers";
-import { FormField, ZacComposedForm } from "../../../shared/form/composed-form/composed-form.component";
+import { ZacComposedForm } from "../../../shared/form/composed-form/composed-form.component";
+import { FormField } from "../../../shared/form/composed-form/form-field.types";
 import { MaterialFormBuilderModule } from "../../../shared/material-form-builder/material-form-builder.module";
 import { GeneratedType } from "../../../shared/utils/generated-types";
 import { AbstractTaskForm } from "./abstract-task-form";
