+++ src/main/app/src/app/formulieren/taken/model/goedkeuren-task-form.ts
@@ -8,7 +8,7 @@ import { Validators } from "@angular/forms";
 import { lastValueFrom } from "rxjs";
 import { mapStringToDocumentenStrings } from "../../../documenten/document-utils";
 import { InformatieObjectenService } from "../../../informatie-objecten/informatie-objecten.service";
-import { FormField } from "../../../shared/form/composed-form/composed-form.component";
+import { FormField } from "../../../shared/form/composed-form/form-field.types";
 import { GeneratedType } from "../../../shared/utils/generated-types";
 import { Goedkeuring } from "../goedkeuring.enum";
 import { AbstractTaskForm } from "./abstract-task-form";
