+++ src/main/app/src/app/shared/form/composed-form/composed-form.component.spec.ts
@@ -17,7 +17,8 @@ import { ZacAutoComplete } from "../auto-complete/auto-complete";
 import { ZacCheckbox } from "../checkbox/checkbox";
 import { ZacDate } from "../date/date";
 import { ZacDocuments } from "../documents/documents";
-import { ZacComposedForm, FormConfig, FormField } from "./composed-form.component";
+import { ZacComposedForm } from "./composed-form.component";
+import { FormConfig, FormField } from "./form-field.types";
 import { ZacHtmlEditor } from "../html-editor/html-editor";
 import { ZacInput } from "../input/input";
 import { ZacRadio } from "../radio/radio";
