+++ src/main/app/src/app/shared/form/composed-form/composed-form.component.ts
@@ -10,140 +10,23 @@ import {
   input,
   output,
 } from "@angular/core";
-import { AbstractControl, FormGroup } from "@angular/forms";
-import { Observable } from "rxjs";
-import { GeneratedType } from "../../utils/generated-types";
-
-type _Form = Record<string, AbstractControl<unknown, unknown>>;
-
-type SingleInputFormField<
-  Form extends _Form,
-  Key extends keyof Form = keyof Form,
-> = {
-  key: Key;
-  label?: string;
-  readonly?: boolean;
-  hidden?: boolean;
-  control?: AbstractControl<unknown, unknown>;
-};
-
-type MultipleInputFormField<
-  Form extends _Form,
-  Key extends keyof Form = keyof Form,
-  Option extends string | Record<string, unknown> = Record<string, unknown>,
-> = SingleInputFormField<Form, Key> & {
-  options: Option[] | Observable<Option[]>;
-  optionDisplayValue?: keyof Option | ((option: Option) => string);
-};
-
-type SelectFormField<
-  Form extends _Form,
-  Key extends keyof Form = keyof Form,
-  Option extends string | Record<string, unknown> = Record<string, unknown>,
-> = MultipleInputFormField<Form, Key, Option> & {
-  type: "select";
-};
-
-type InputFormField<Form extends _Form> = SingleInputFormField<Form> & {
-  type: "input";
-};
-
-type TextareaFormField<Form extends _Form> = SingleInputFormField<Form> & {
-  type: "textarea";
-};
-
-type DateFormField<Form extends _Form> = SingleInputFormField<Form> & {
-  type: "date";
-};
-
-type HtmlEditorField<Form extends _Form> = SingleInputFormField<Form> & {
-  type: "html-editor";
-  variables?: string[];
-};
-
-type CheckboxField<Form extends _Form> = SingleInputFormField<Form> & {
-  type: "checkbox";
-};
-
-type AutocompleteFormField<
-  Form extends _Form,
-  Key extends keyof Form = keyof Form,
-  Option extends string | Record<string, unknown> = Record<string, unknown>,
-> = MultipleInputFormField<Form, Key, Option> & {
-  type: "auto-complete";
-};
-
-type DocumentFormField<
-  Form extends _Form,
-  Key extends keyof Form = keyof Form,
-  Option extends
-    GeneratedType<"RestEnkelvoudigInformatieobject"> = GeneratedType<"RestEnkelvoudigInformatieobject">,
-> = MultipleInputFormField<Form, Key, Option> & {
-  type: "documents";
-  viewDocumentInNewTab?: boolean;
-};
-
-type PlainTextField<Form extends _Form> = SingleInputFormField<Form> & {
-  type: "plain-text";
-  icon?: string;
-};
-
-type RadioFormField<
-  Form extends _Form,
-  Key extends keyof Form = keyof Form,
-  Option extends string | Record<string, unknown> =
-    | string
-    | Record<string, unknown>,
-> = MultipleInputFormField<Form, Key, Option> & {
-  type: "radio";
-};
-
-/**
- * This type is meant to be used **only** in the `ZacComposedForm` component.
- */
-type _FormConfig = {
-  submitLabel?: string;
-};
-type FormConfigWithPartialSubmit = _FormConfig & {
-  partialSubmitLabel: string;
-  hideCancelButton: true;
-  cancelLabel?: null;
-};
-
-type CancelableFormConfig = _FormConfig & {
-  partialSubmitLabel?: null;
-  cancelLabel?: string;
-  hideCancelButton?: false;
-};
-
-export type FormConfig = FormConfigWithPartialSubmit | CancelableFormConfig;
-
-export type FormField<Form extends _Form = _Form> =
-  | SelectFormField<Form>
-  | InputFormField<Form>
-  | DateFormField<Form>
-  | HtmlEditorField<Form>
-  | CheckboxField<Form>
-  | TextareaFormField<Form>
-  | AutocompleteFormField<Form>
-  | DocumentFormField<Form>
-  | PlainTextField<Form>
-  | RadioFormField<Form>;
+import { FormGroup } from "@angular/forms";
+import { Form, FormConfig, FormField } from "./form-field.types";
 
 @Component({
   selector: "zac-composed-form",
   templateUrl: "./composed-form.component.html",
   standalone: false,
 })
-export class ZacComposedForm<Form extends _Form> {
-  protected readonly form = input.required<FormGroup<Form>>();
+export class ZacComposedForm<F extends Form> {
+  protected readonly form = input.required<FormGroup<F>>();
   protected readonly fields = input.required<FormField[]>();
   protected readonly config = input<FormConfig>({ hideCancelButton: false });
   protected readonly readonly = input(false, { transform: booleanAttribute });
   protected readonly loading = input(false, { transform: booleanAttribute });
 
-  protected readonly formSubmitted = output<FormGroup<Form>>();
-  protected readonly formPartiallySubmitted = output<FormGroup<Form>>();
+  protected readonly formSubmitted = output<FormGroup<F>>();
+  protected readonly formPartiallySubmitted = output<FormGroup<F>>();
   protected readonly formCancelled = output<void>();
 
   constructor() {
